requirements.txt
PySide6
Pillow
rembg
numpy
